Improve the lit renderer structure
