+++ renderers/lit/src/0.8/core/standard_catalog_api/text.ts
@@ -7,7 +7,8 @@
 
       https://www.apache.org/licenses/LICENSE-2.0
 
- Unless required by applicable law of an "AS IS" BASIS,
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  See the License for the specific language governing permissions and
  limitations under the License.
@@ -15,9 +16,16 @@
 
 import {
   ComponentApi,
-  TextNode,
+  BaseResolvedNode,
 } from '../types/types.js';
 import { StringValue } from '../types/primitives.js';
+
+export interface TextNode extends BaseResolvedNode<'Text'> {
+  properties: {
+    text: StringValue;
+    usageHint: "h1" | "h2" | "h3" | "h4" | "h5" | "caption" | "body";
+  }
+}
 
 export const textApi: ComponentApi<'Text', TextNode> = {
   name: 'Text',
