+++ renderers/lit/src/0.8/lit/components/divider.ts
@@ -20,7 +20,7 @@ import { Root } from "./root.js";
 import { classMap } from "lit/directives/class-map.js";
 import { styleMap } from "lit/directives/style-map.js";
 import { structuralStyles } from "./styles.js";
-import { DividerNode } from "../../core/types/types.js";
+import { DividerNode } from "../../core/standard_catalog_api/divider.js";
 
 @customElement("a2ui-divider")
 export class Divider extends Root<DividerNode> {
