+++ renderers/lit/src/0.8/lit/components/icon.ts
@@ -21,7 +21,7 @@ import { classMap } from "lit/directives/class-map.js";
 import { A2uiMessageProcessor } from "../../core/a2ui_message_processor.js";
 import { styleMap } from "lit/directives/style-map.js";
 import { structuralStyles } from "./styles.js";
-import { IconNode } from "../../core/types/types.js";
+import { IconNode } from "../../core/standard_catalog_api/icon.js";
 
 @customElement("a2ui-icon")
 export class Icon extends Root<IconNode> {
