+++ renderers/lit/src/0.8/lit/components/row.ts
@@ -18,7 +18,7 @@ import { html, css, nothing } from "lit";
 import { customElement, property } from "lit/decorators.js";
 import { Root } from "./root.js";
 import { classMap } from "lit/directives/class-map.js";
-import { RowNode } from "../../core/types/types.js";
+import { RowNode } from "../../core/standard_catalog_api/row.js";
 import { styleMap } from "lit/directives/style-map.js";
 import { structuralStyles } from "./styles.js";
