+++ renderers/lit/src/0.8/lit/components/video.ts
@@ -20,7 +20,7 @@ import { Root } from "./root.js";
 import { classMap } from "lit/directives/class-map.js";
 import { styleMap } from "lit/directives/style-map.js";
 import { structuralStyles } from "./styles.js";
-import { VideoNode } from "../../core/types/types.js";
+import { VideoNode } from "../../core/standard_catalog_api/video.js";
 import { A2uiMessageProcessor } from "../../core/a2ui_message_processor.js";
 
 @customElement("a2ui-video")
