+++ renderers/lit/src/0.8/lit/standard_catalog_implementation/datetime_input.ts
@@ -16,7 +16,7 @@
 
 import { html, TemplateResult } from 'lit';
 import { ComponentRenderer } from '../../core/types/types.js';
-import { DateTimeInputNode } from '../../core/types/types.js';
+import { DateTimeInputNode } from '../../core/standard_catalog_api/datetime_input.js';
 import '../components/datetime-input.js';
 
 export const litDateTimeInputRenderer: ComponentRenderer<DateTimeInputNode, TemplateResult> = {
